refactor: Move form components to general forms directory
- Move all form components from WorldEditor to components/forms
- Update Storybook hierarchy to show forms in their own section
- Update all imports to reference new location
- Maintain all functionality while making forms more reusable
